Module profile::heartbeats

Statics

HBS

Functions

cleanup
Log regmaining buffer data and cleanup heartbeats

get_heartbeat_log

get_heartbeat_window_size

heartbeat_window_callback
Callback function used to log the window buffer. When this is called from native C, the heartbeat is safely locked

init
Initialize heartbeats

is_create_heartbeat

is_heartbeat_enabled
Check if a heartbeat exists for the given category

log_heartbeat_records
Log heartbeat records up to the buffer index

maybe_create_heartbeat
Possibly create a heartbeat

maybe_heartbeat
Issue a heartbeat (if one exists) for the given category

open_heartbeat_log