Enum profile::mem::ProfilerMsg

pub enum ProfilerMsg {
    RegisterReporter(String, Box<Reporter + Send>),
    UnregisterReporter(String),
    Print,
    Exit,
}

Messages that can be sent to the memory profiler thread.

Variants

RegisterReporter

Register a Reporter with the memory profiler. The String is only used to identify the reporter so it can be unregistered later. The String must be distinct from that used by any other registered reporter otherwise a panic will occur.

UnregisterReporter

Unregister a Reporter with the memory profiler. The String must match the name given when the reporter was registered. If the String does not match the name of a registered reporter a panic will occur.

Print

Triggers printing of the memory profiling metrics.

Exit

Tells the memory profiler to shut down.