Module profile::mem

Memory profiling functions.

Modules

system_reporter

Structs

Profiler

ReportsForest

A collection of ReportsTrees. It represents the data from multiple memory reports in a form that’s good to print.

ReportsTree

A collection of one or more reports with the same initial path segment. A ReportsTree containing a single node is described as “degenerate”.

Constants

JEMALLOC_HEAP_ALLOCATED_STR

SYSTEM_HEAP_ALLOCATED_STR