Module profile::mem

Memory profiling functions.

Structs

Profiler

ProfilerChan

Report
A single memory-related measurement.

ReportsChan
A channel through which memory reports can be sent.

Enums

ProfilerMsg
Messages that can be sent to the memory profiler thread.

Traits

Reporter
A memory reporter is capable of measuring some data structure of interest. Because it needs to be passed to and registered with the Profiler, it's typically a "small" (i.e. easily cloneable) value that provides access to a "large" data structure, e.g. a channel that can inject a request for measurements into the event queue associated with the "large" data structure.